refactor: middleware and nonmiddleware share the same module
